Convert GDP data

convertGDP() converts GDP time series data from one unit to another, using GDP deflators, market exchange rates (MERs) and purchasing power parity conversion factors (PPPs).

convertGDP(
  gdp,
  unit_in,
  unit_out,
  source = "wb_wdi",
  with_regions = NULL,
  replace_NAs = NULL,
  verbose = FALSE,
  return_cfs = FALSE
)

Arguments

gdp

A tibble, data frame or magpie object, the latter of which requires the magclass package to be installed. The data-frame needs to have at least 3 columns:

a character column with iso3c (wikipedia) country codes,

a numeric column with years,

a numeric column named "value" with GDP values.

unit_in

A string with the incoming GDP unit, one of:

"current LCU"

"current Int$PPP"

"current US$MER"

"constant YYYY LCU"

"constant YYYY Int$PPP"

"constant YYYY US$MER"

where YYYY should be replaced with a year e.g. "2010" or "2017".

unit_out

A string with the outgoing GDP unit, one of:

"current LCU"

"current Int$PPP"

"current US$MER"

"constant YYYY LCU"

"constant YYYY Int$PPP"

"constant YYYY US$MER"

where YYYY should be replaced with a year e.g. "2010" or "2017".

source

A data frame, or the name of a data frame, to use for conversion factors. Can be a custom data-frame that exists in the calling environment, or one of the package internal ones. Use print_source_info() to learn about the available sources.

with_regions

NULL or a data-frame. The data-frame should be "country to region mapping": one column named "iso3c" with iso3c country codes, and one column named "region" with region codes to which the countries belong. Any regions in the gdp object will then be disaggregated according to the region mapping and weighed by the GDP share of countries in that region in the year of the unit, converted on a country level, and re-aggregated before being returned.

replace_NAs

NULL, 0, 1 or "regional_average". Should countries for which conversion factors are missing, have their NA-conversion factors replaced with 1 or with a regional average? The default is no. If 0, then resulting NAs are simply replaced with 0. If 1, then the missing conversion factors will be set to 1, and essentially, no conversion will take place. If "regional_average" then, the regional average of the region to which the country belongs to will be used. This requires a region-mapping to be passed to the function, see the with_regions argument.

verbose

TRUE or FALSE. A flag to turn verbosity on or off. Overrules the GDPuc.verbose option, if it is set.

return_cfs

TRUE or FALSE. Set to TRUE to additionally return a tibble with the conversion factors used. In that case a list is returned with the converted GDP under "result", and the conversion factors used under "cfs".

Value

The gdp argument, with the values in the "value" column, converted to unit_out. If the argument return_cfs is TRUE, then a list is returned with the converted GDP under "result", and the conversion factors used under "cfs".

Details

When providing a custom source to the function, a certain format is required. The source object must be a data frame or tibble with at least the following columns:

a character column named "iso3c" with iso3c (wikipedia) country codes,

a numeric column named "year" with years,

a numeric column named "GDP deflator" with values of the GDP deflator divided by 100 (so that in the base year the GDP deflator is equal to 1, not 100). The base year of the deflator can be any year, and can be country-specific.

a numeric column named "MER (LCU per US$)" with MER values,

a numeric column named "PPP conversion factor, GDP (LCU per international $)" wit PPP exchange rate values.

See also